Separate components into categoric folders in an attempt to keep codebase clean. Create css files for each specific component.
