+++ react-app/src/App.css
@@ -1,383 +1,3 @@
 html, body {
 	overflow-x: hidden;
 }
-
-@media screen 
-  and (min-device-width: 1200px)  
-  and (-webkit-min-device-pixel-ratio: 1) {  
-
-	div {
-		width: 100%;
-	}
-
-	#title-logo {
-		text-align: center;
-		max-width: 100%;
-		max-height: 100%;
-	}
-
-	#logo-div {
-		text-align: center;
-	}
-
-	#social {
-		margin: 0 auto;
-		text-align: center;
-		padding-right: 60px;
-		padding-bottom: 50px;
-	}
-
-	li {
-		display: inline-block;
-	}
-
-	#twitter{
-		padding: 15%;
-	}
-
-	#instagram {
-		padding: 15%;
-	}
-
-	#facebook {
-		padding: 15%;
-	}
-
-	#twitter img {
-		width: 3.5vw;
-	}
-
-	#instagram img  {
-		width: 3.5vw;
-	}
-
-	#facebook img  {
-		width: 3.5vw;
-	}
-}
-
-@media screen 
-  and (max-device-width: 1200px)  
-  and (-webkit-min-device-pixel-ratio: 2) {  
-
-	div {
-		width: 100%;
-	}
-
-	#title-logo {
-		text-align: center;
-		max-width: 100%;
-		max-height: 100%;
-	}
-
-	#logo-div {
-		text-align: center;
-	}
-
-	#social {
-		margin: 0 auto;
-		text-align: center;
-		padding-right: 60px;
-		padding-bottom: 50px;
-	}
-
-	li {
-		display: inline-block;
-		text-decoration: none;
-	}
-
-	#twitter{
-		padding: 5.5%;
-	}
-
-	#instagram {
-		padding: 5.5%;
-	}
-
-	#facebook {
-		padding: 5.5%;
-	}
-
-	#twitter img {
-		width: 9.5vw;
-	}
-
-	#instagram img  {
-		width: 9.5vw;
-	}
-
-	#facebook img  {
-		width: 9.5vw;
-	}
-}
-
-@media screen 
-  and (min-device-width: 1200px)  
-  and (-webkit-min-device-pixel-ratio: 1) { 
-
-	@font-face {
-		font-family: AvenirNext;
-		src: url("res/fonts/AvenirNextLTPro-Regular.otf") format("opentype");
-	}
-
-	#eo-div {
-		background-color: rgba(35, 36, 35, 1.0);
-		position: relative;
-		height: auto;
-		right: 0%;
-		width: 100%;
-	}
-
-	#eo-plate {
-		margin-top: 1.5%;
-		position: relative;
-		text-align: left;
-	  	width: 80%;
-	  	min-height: 50%;
-	}
-
-	#eo-plate img {
-		width: 60%;
-	}
-
-	#eo-ios {
-		margin-top: -46%;
-		margin-left: 11.5%;
-		position: relative;
-		text-align: left;
-		width: 70%;
-		height: 70%;	
-	}
-
-	#eo-ios img {
-		width: 35%;	
-	}
-
-	#eo-content {
-		font-family: AvenirNext;
-		margin-top: -35%;
-		text-align: right;
-		color: white;
-	}
-
-	#eo-logo {
-		max-height: 50%;
-	}
-
-	#eo-logo img {
-		margin-right: 7%;
-		max-height: 35%;
-		max-width: 35%;
-	}
-
-	#eo-info, #eo-motto {
-		color: white;
-		width: 35%;
-		margin-left: 57%;
-		text-align: center;
-	}
-
-	#eo-motto {
-		font-size: 4vw;
-		animation: motto-slide 1.75s;
-		-webkit-animation: motto-slide 1.75s;	
-	}
-
-	#eo-info {
-		font-size: 2vw;
-		animation: info-fade 3s;
-	}
-}
-
-@media screen 
-  and (max-device-width: 1200px)  
-  and (-webkit-min-device-pixel-ratio: 2) { 
-
-	@font-face {
-		font-family: AvenirNext;
-		src: url("res/fonts/AvenirNextLTPro-Regular.otf") format("opentype");
-	}
-
-	#eo-div {
-		background-color: rgba(35, 36, 35, 1.0);
-		position: relative;
-		height: auto;
-		right: 0%;
-		width: 100%;
-	}
-
-	#eo-plate {
-		padding-top: 56%;
-		margin-top: 1.5%;
-		position: relative;
-		text-align: center;
-	  	width: 100%;
-	  	min-height: 50%;
-	  	bottom: 0px;
-	}
-
-	#eo-plate img {
-		width: 60%;
-	}
-
-	#eo-ios {
-		margin-top: -62%;
-		position: relative;
-		text-align: center;
-		width: 100%;
-		height: 70%;	
-	}
-
-	#eo-ios img {
-		width: 35%;	
-	}
-
-	#eo-content {
-		font-family: AvenirNext;
-		margin-top: -134%;
-		text-align: center;
-		color: white;
-	}
-
-	#eo-logo {
-		max-height: 50%;
-	}
-
-	#eo-logo img {
-		max-height: 35%;
-		max-width: 35%;
-	}
-
-	#eo-info, #eo-motto {
-		color: white;
-		width: 55%;
-		text-align: center;
-		margin-left: 22%;
-	}
-
-	#eo-motto {
-		font-size: 6vw;
-		animation: motto-slide 1.75s;
-		-webkit-animation: motto-slide 1.75s;	
-	}
-
-	#eo-info {
-		font-size: 4.5vw;
-		animation: info-fade 3s;
-	}
-}
-
-@media screen 
-  and (min-device-width: 1200px)  
-  and (-webkit-min-device-pixel-ratio: 1) { 
-
-	#about-div {
-		min-height: 500px;
-		position: relative;
-		text-align: center;
-	}
-
-	#about-title {
-		font-family: AvenirNext;
-		font-size: 3vw;
-	}
-
-	#about-join {
-		font-family: AvenirNext;
-		font-size: 1.5vw;	
-		color: black;
-	}
-
-	.profile_pics {
-		position: relative;
-		text-align: center;
-		font-family: AvenirNext;
-		padding: 0% 2.5% 0% 2.5%;
-	}
-
-	a {
-		text-decoration: none;
-		color: black;
-	}
-
-	a:hover {
-		text-decoration: underline;
-	}
-
-	.profile_pics .portrait_name {
-		font-size: 2vw;
-	}
-
-	.profile_pics .portrait_description {
-		font-size: 1vw;
-	}
-
-	.profile_pics img {
-		margin-top: 15%;
-		height: 12vw;
-		border-radius: 6vw;
-		zoom: 125%;
-	}
-	.gif img{
-		width: 25px;
-		height: auto;
-	}
-}
-
-@media screen 
-  and (max-device-width: 1200px)  
-  and (-webkit-min-device-pixel-ratio: 2) { 
-
-  	li {
-  		display: inline;
-  	}
-	#about-div {
-		min-height: 500px;
-		position: relative;
-		text-align: center;
-	}
-
-	#about-title {
-		font-family: AvenirNext;
-		font-size: 8vw;
-	}
-
-	#about-join {
-		font-family: AvenirNext;
-		font-size: 3.5vw;	
-		color: black;
-	}
-
-	a {
-		text-decoration: underline;
-		color: black;
-	}
-
-	.profile_pics {
-		position: relative;
-		text-align: center;
-		font-family: AvenirNext;
-	}
-
-	.profile_pics .portrait_name {
-		font-size: 6vw;
-		margin-right: 3.5%;
-	}
-
-	.profile_pics .portrait_description {
-		font-size: 3.5vw;
-		margin-right: 3.5%;
-	}
-
-	.profile_pics img {
-		margin-right: 3.5%;
-		margin-top: 15%;
-		height: 40vw;
-		border-radius: 15vw;
-		zoom: 125%;
-	}
-	.gif img{
-		width: 25px;
-		height: auto;
-	}
-}
