--- react-app/src/EatOut.css
@@ -1,7 +1,0 @@
-.EatOut {
-    text-decoration: none;
-}
-
-.EatOut:hover {
-    text-decoration: none;
-}
