+++ react-app/src/EatOut/EatOut.js
@@ -1,8 +1,8 @@
 import React, { Component } from 'react'
 
-import EatOutLogo from './res/images/Eat Out Logo.png';
-import iOSMockup from './res/images/iOS-mockup.png'
-import Plate from './res/images/plate.png'
+import EatOutLogo from '../res/images/Eat Out Logo.png';
+import iOSMockup from '../res/images/iOS-mockup.png'
+import Plate from '../res/images/plate.png'
 
 import './EatOut.css'
